Development

+++ BiBoGC/InventoryManagement.Application/Commands/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -22,7 +22,20 @@ namespace InventoryManagement.Application.Commands.DeleteCategory
                 return Result<Guid>.Failure($"Danh mục có mã {request.Id} không tồn tại");
             }
 
+            // adding logic for cascading delete
+
+            var subCategories = await _categoryRepository.GetSubCategoriesAsync(request.Id);
+
+            if(subCategories.Count() > 0)
+            {
+                return Result<Guid>.Failure($"Danh mục này có danh mục con, hãy xóa danh mục con trước.");
+            }
+
+
             await _categoryRepository.DeleteAsync(category, cancellationToken);
+
+
+            
             return Result<Guid>.Success(request.Id);
         }
     }
